Use a build script instead of abusing unit tests
