--- src/consts/values.rs
@@ -1,66 +1,0 @@
-//! Predefined parameter values, @generated in tests/codegen.rs
-//!
-//! # Sources
-//! - <https://www.iana.org/assignments/character-sets/character-sets.xhtml>
-//! - <https://datatracker.ietf.org/doc/html/rfc3676>
-
-/// `Big5`
-pub const BIG5: crate::Value = crate::Value::new_unchecked("Big5");
-/// `EUC-JP`
-pub const EUC_JP: crate::Value = crate::Value::new_unchecked("EUC-JP");
-/// `EUC-KR`
-pub const EUC_KR: crate::Value = crate::Value::new_unchecked("EUC-KR");
-/// `fixed`
-pub const FIXED: crate::Value = crate::Value::new_unchecked("fixed");
-/// `flowed`
-pub const FLOWED: crate::Value = crate::Value::new_unchecked("flowed");
-/// `GB2312`
-pub const GB2312: crate::Value = crate::Value::new_unchecked("GB2312");
-/// `ISO-2022-JP`
-pub const ISO_2022_JP: crate::Value = crate::Value::new_unchecked("ISO-2022-JP");
-/// `ISO-2022-JP-2`
-pub const ISO_2022_JP_2: crate::Value = crate::Value::new_unchecked("ISO-2022-JP-2");
-/// `ISO-2022-KR`
-pub const ISO_2022_KR: crate::Value = crate::Value::new_unchecked("ISO-2022-KR");
-/// `ISO-8859-1`
-pub const ISO_8859_1: crate::Value = crate::Value::new_unchecked("ISO-8859-1");
-/// `ISO-8859-10`
-pub const ISO_8859_10: crate::Value = crate::Value::new_unchecked("ISO-8859-10");
-/// `ISO-8859-2`
-pub const ISO_8859_2: crate::Value = crate::Value::new_unchecked("ISO-8859-2");
-/// `ISO-8859-3`
-pub const ISO_8859_3: crate::Value = crate::Value::new_unchecked("ISO-8859-3");
-/// `ISO-8859-4`
-pub const ISO_8859_4: crate::Value = crate::Value::new_unchecked("ISO-8859-4");
-/// `ISO-8859-5`
-pub const ISO_8859_5: crate::Value = crate::Value::new_unchecked("ISO-8859-5");
-/// `ISO-8859-6`
-pub const ISO_8859_6: crate::Value = crate::Value::new_unchecked("ISO-8859-6");
-/// `ISO-8859-6-E`
-pub const ISO_8859_6_E: crate::Value = crate::Value::new_unchecked("ISO-8859-6-E");
-/// `ISO-8859-6-I`
-pub const ISO_8859_6_I: crate::Value = crate::Value::new_unchecked("ISO-8859-6-I");
-/// `ISO-8859-7`
-pub const ISO_8859_7: crate::Value = crate::Value::new_unchecked("ISO-8859-7");
-/// `ISO-8859-8`
-pub const ISO_8859_8: crate::Value = crate::Value::new_unchecked("ISO-8859-8");
-/// `ISO-8859-8-E`
-pub const ISO_8859_8_E: crate::Value = crate::Value::new_unchecked("ISO-8859-8-E");
-/// `ISO-8859-8-I`
-pub const ISO_8859_8_I: crate::Value = crate::Value::new_unchecked("ISO-8859-8-I");
-/// `ISO-8859-9`
-pub const ISO_8859_9: crate::Value = crate::Value::new_unchecked("ISO-8859-9");
-/// `KOI8-R`
-pub const KOI8_R: crate::Value = crate::Value::new_unchecked("KOI8-R");
-/// `no`
-pub const NO: crate::Value = crate::Value::new_unchecked("no");
-/// `Shift_JIS`
-pub const SHIFT_JIS: crate::Value = crate::Value::new_unchecked("Shift_JIS");
-/// `US-ASCII`
-pub const US_ASCII: crate::Value = crate::Value::new_unchecked("US-ASCII");
-/// `UTF-8`
-pub const UTF_8: crate::Value = crate::Value::new_unchecked("UTF-8");
-/// `UTF-16`
-pub const UTF_16: crate::Value = crate::Value::new_unchecked("UTF-16");
-/// `yes`
-pub const YES: crate::Value = crate::Value::new_unchecked("yes");
